reorg of files to be under conversion when specific to that
