+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/OpFusionHelper.hpp
@@ -99,15 +99,15 @@ public:
 // Subclass for ONNXFusedOp(kind = "zhigh.expand-mul-stick").
 //
 // Pattern:
-//   ONNXUnsqueezeOp  one axis P; innermost dim of result static mod 64
-//                    (required)
-//  ONNXExpandOp      dim P expands from 1 to N (N static, >= 2)
-//                    (required)
-// ONNXMulOp          element-wise mul by scalar F32/I32/I64 const
-//                    (optional; when absent, mulScalar stays at its neutral
+//  ONNXUnsqueezeOp  one axis P; innermost dim of result static mod 64
+//                   (required)
+//  ONNXExpandOp     dim P expands from 1 to N (N static, >= 2)
+//                   (required)
+//  ONNXMulOp        element-wise mul by scalar F32/I32/I64 const
+//                   (optional; when absent, mulScalar stays at its neutral
 //                    1.f default)
-//   ONNXReshapeOp    dims 0..P may collapse; dims after P unchanged (required)
-//   ZHighStickOp     stick to 3D / 3DS / 4D (required)
+//  ONNXReshapeOp    dims 0..P may collapse; dims after P unchanged (required)
+//  ZHighStickOp     stick to 3D / 3DS / 4D (required)
 //
 // Unique-use invariant: every intermediate value (unsqueeze through reshape)
 // has exactly one use.  The stick result is not checked.
