+++ src/Conversion/ONNXToKrnl/ONNXToKrnlCommon.cpp
@@ -1113,4 +1113,16 @@ void genSafeCodeForGatherAlike(mlir::ConversionPatternRewriter &rewriter,
       });
 }
 
+//===----------------------------------------------------------------------===//
+// FusedOpInlineFallback
+//===----------------------------------------------------------------------===//
+
+LogicalResult FusedOpInlineFallback::matchAndRewrite(
+    ONNXFusedOp fusedOp, OpAdaptor, ConversionPatternRewriter &rewriter) const {
+  fusedOp.emitWarning()
+      << "no dedicated lowering for onnx.Fused (kind='" << fusedOp.getKind()
+      << "'); inlining body as fallback — add a FusedOpKindLowering subclass";
+  return FusionOpChain::inlineFallback(rewriter, fusedOp);
+}
+
 } // namespace onnx_mlir
